2.9.10 - Исправлена ошибка вычисления сигнатур файлов логов - Оптимизировано чтение входящих POST-параметров - Добавлено свойство "bm" в записи логов - битовая закладка в файле

+++ src/main/java/com/monitor/parser/reader/ParserFileReader.java
@@ -83,7 +83,7 @@ public class ParserFileReader {
         }
         if (!draft) {
             for (Entry<FileState, Long> entry : pointerMap.entrySet()) {
-                entry.getKey().setNewPointer(entry.getValue().longValue());
+                entry.getKey().setNewPointer(entry.getValue());
             }
             
 //          for (FileState state : fileList) {
